renamed misc.py and qd_functions.py to, respectively, utils.py and mol_utils.py
